backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
django-stubs==6.0.3
django-stubs-ext==6.0.3
django-types==0.24.0
djangorestframework==3.17.1
djangorestframework-stubs==3.16.9
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
PyJWT==2.12.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
types-psycopg2==2.9.21.20260422
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
uritemplate==4.2.0
